Refactored metrics calculation code
